+++ vcm/downloader/link.py
@@ -652,6 +652,14 @@ class Quiz(BaseUndownloableLink):
 
     NOTIFY = True
 
+class BlackBoard(BaseUndownloableLink):
+    """Representation of a blackboard link.
+
+    A blackboard is a link to a VoIP chat.
+    """
+
+    NOTIFY = True
+
 
 class Html(BaseLink):
     def do_download(self):